Moves are prepared and MoveSelection scene
Move selection scene - this scene has the moves on it and the battle options class is reworked so now that the fight button take you to the move selection scene and the battle method knows what move is being done.

move class - this class has all of the info of what the move is. Such as the BP, Type, and Stat being used.

other classes - have been updated to accommodate the move class and the moveSelection scene.